personal-site/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-grappelli==2.14.2
django-tinymce==3.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
image==1.5.33
install==1.3.4
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
